+++ src/main/java/io/personium/core/model/ctl/Rule.java
@@ -1,6 +1,6 @@
 /**
  * personium.io
- * Copyright 2017 FUJITSU LIMITED
+ * Copyright 2017-2018 FUJITSU LIMITED
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -45,6 +45,8 @@ public class Rule {
     public static final String ACTION_EXEC = "exec";
     /** action relay. */
     public static final String ACTION_RELAY = "relay";
+    /** action relay.event. */
+    public static final String ACTION_RELAY_EVENT = "relay.event";
     /** action log. */
     public static final String ACTION_LOG = "log";
     /** action log.info. */
@@ -60,7 +62,7 @@ public class Rule {
     public static final String P_FORMAT_PATTERN_RULE_SERVICE = "rule-service";
 
     /** Pattern action. */
-    private static final String PATTERN_ACTION = "^(exec)|(relay)|(log)|(log\\.info)|(log\\.warn)|(log\\.error)$"; // CHECKSTYLE IGNORE - To maintein readability
+    private static final String PATTERN_ACTION = "^(exec)|(relay)|(relay\\.event)|(log)|(log\\.info)|(log\\.warn)|(log\\.error)$"; // CHECKSTYLE IGNORE - To maintein readability
 
     /** Annotations for Object. */
     private static final List<EdmAnnotation<?>> P_FORMAT_OBJECT = new ArrayList<EdmAnnotation<?>>();