+++ src/main/java/io/personium/core/rs/cell/CellResource.java
@@ -71,6 +71,7 @@ public final class CellResource {
      * @param requestKey X-Personium-RequestKey header
      * @param eventId X-Personium-EventId header
      * @param ruleChain X-Personium-RuleChain header
+     * @param via X-Personium-Via header
      * @param httpServletRequest HttpServletRequest
      */
     public CellResource(
@@ -78,6 +79,7 @@ public final class CellResource {
             final String requestKey,
             final String eventId,
             final String ruleChain,
+            final String via,
             HttpServletRequest httpServletRequest) {
         // Cellが存在しないときは例外
         this.accessContext = accessContext;
@@ -91,7 +93,8 @@ public final class CellResource {
             throw PersoniumCoreException.Dav.CELL_NOT_FOUND;
         }
 
-        this.cellRsCmp = new CellRsCmp(this.cellCmp, this.cell, this.accessContext, requestKey, eventId, ruleChain);
+        this.cellRsCmp = new CellRsCmp(this.cellCmp, this.cell, this.accessContext,
+                requestKey, eventId, ruleChain, via);
         checkReferenceMode();
 
         // If cell status is import failed, APIs other than import or token or BulkDeletion are not accepted.