+++ src/main/java/io/personium/core/rule/EventReceiver.java
@@ -69,8 +69,10 @@ class EventReceiver implements Runnable {
                         log.debug("    RequestKey: " + event.getRequestKey());
                         log.debug("    EventId: " + event.getEventId());
                         log.debug("    RuleChain: " + event.getRuleChain());
+                        log.debug("    Via: " + event.getVia());
+                        log.debug("    Roles: " + event.getRoles());
                         log.debug("    CellId: " + event.getCellId());
-                        log.debug("    DateTime: " + event.getDateTime());
+                        log.debug("    Time: " + event.getTime());
 
                         // If the event matches with a rule, execute action of the rule.
                         rman.judge(event);