+++ src/main/java/io/personium/core/ws/EventSubscriber.java
@@ -65,8 +65,9 @@ public class EventSubscriber implements Runnable {
                         log.debug("    RequestKey: " + event.getRequestKey());
                         log.debug("    EventId: " + event.getEventId());
                         log.debug("    RuleChain: " + event.getRuleChain());
+                        log.debug("    Via: " + event.getVia());
                         log.debug("    CellId: " + event.getCellId());
-                        log.debug("    DateTime: " + event.getDateTime());
+                        log.debug("    Time: " + event.getTime());
 
                         // WebSocket send
                         WebSocketService.sendEvent(event);